+++ doc/CHANGELOG.md
@@ -8,9 +8,10 @@
 * `#1789` Move validation to Work Package
 * `#1800` Add settings to change software name and URL and add additional footer content
 * `#1808` Add option to log user for each request
-* `#1875` Added test steps to reuse steps for my page, my project page, and documents
+* `#1875` Added test steps to reuse steps for my page, my project page, and documents, no my page block lookup at class load time
 * `#1876` Timelines do not show work packages when there is no status reporting
 * `#1896` Moved visibility-tests for issues into specs for workpackages
+* `#1912` Merge column project type with column planning element type
 
 ## 3.0.0pre14